requirements.txt
pandas
numpy
torch
transformers
scipy
langdetect
deep-translator
matplotlib
seaborn
openpyxl
tabulate
spacy
nbformat
streamlit
plotly
tqdm
scikit-learn
accelerate
sentencepiece
fpdf2
pymongo
google-generativeai
